feat: serve the portfolio at the home route (/)
Move the Sanity-driven portfolio to `/` and retire the legacy home page.

- pages/index.tsx now renders the portfolio (getStaticProps + ISR)
- removed pages/portfolio.tsx
- next.config: permanent redirect /portfolio -> /
- portfolio-report link points at the home page
